[FIX] move include/seqan3/{core,utility}/math.hpp
### move header and test

```bash
mkdir -p include/seqan3/utility # maybe create target folder

cp include/seqan3/{core,utility}/math.hpp
git mv test/unit/{core,utility}/math_test.cpp # if it exists
```

* [x] move header and tests

### check new header

`include/seqan3/utility/math.hpp` has
* [x] no unused headers
* [x] include order is correct (see 4. in https://github.com/seqan/seqan3/wiki/File-structure-and-naming#contents)

```bash
edit include/seqan3/utility/math.hpp
git add include/seqan3/utility/math.hpp
```

### deprecate header

```bash
edit include/seqan3/core/math.hpp
git add include/seqan3/core/math.hpp
```

* [x] in `\file` append `\brief` with [DEPRECATED]
* [x] in `\file` add `\deprecated This header will be removed in 3.1. Please use seqan3/utility/math.hpp instead.`
* [x] delete everything after `#pragma once`
* [x] add include after `#pragma once` to new header `#include <include/seqan3/utility/math.hpp>`
* [x] add deprecation notice after `#include`

  ```c++
  SEQAN3_DEPRECATED_HEADER(
     "This header is deprecated and will be removed in SeqAn-3.1.0; Please #include <seqan3/alphabet/aminoacid/translation_new.hpp> instead.")
  ```

File looks like this:

```c++
// -----------------------------------------------------------------------------------------------------
// Copyright (c) 2006-2020, Knut Reinert & Freie Universität Berlin
// Copyright (c) 2016-2020, Knut Reinert & MPI für molekulare Genetik
// This file may be used, modified and/or redistributed under the terms of the 3-clause BSD-License
// shipped with this file and also available at: https://github.com/seqan/seqan3/blob/master/LICENSE.md
// -----------------------------------------------------------------------------------------------------

/*!\file
 * \brief [DEPRECATED] Provides math related functionality.
 * \author Enrico Seiler <enrico.seiler AT fu-berlin.de>
 * \deprecated This header will be removed in 3.1. Please use seqan3/utility/math.hpp instead.
 */

#pragma once

#include <seqan3/utility/math.hpp>

SEQAN3_DEPRECATED_HEADER(
    "This header is deprecated and will be removed in SeqAn-3.1.0; Please #include <seqan3/utility/math.hpp> instead.")
```

### search and replace includes

* [x] #include <seqan3/core/math.hpp> by #include <seqan3/utility/math.hpp>

```bash
git add -p # double check if change was correct
```

### edit CMakeLists.txt

```bash
edit test/unit/core/CMakeLists.txt # remove seqan3_test entry
edit test/unit/utility/CMakeLists.txt # add seqan3_test entry
git add test/unit/core/CMakeLists.txt
git add test/unit/utility/CMakeLists.txt
```

### update snippets

* [x] go through include/seqan3/utility/math.hpp (search for `\include` or `\snippet`)

In this case I got the following snippets
* `test/snippet/core/detail/floor_log2.cpp`
* `test/snippet/core/detail/ceil_log2.cpp`
* `test/snippet/core/pow.cpp`

```bash
mkdir -p test/snippet/utility
mkdir -p test/snippet/utility/detail
git mv test/snippet/{core,utility}/detail/floor_log2.cpp
git mv test/snippet/{core,utility}/detail/ceil_log2.cpp
git mv test/snippet/{core,utility}/pow.cpp

edit include/seqan3/utility/math.hpp # change snippets paths
git add -p
```

### update performance tests

* [x] update `test/unit/performance/`
  * (we only have performance tests for char_operations, simd and bit_manipulation_benchmark.cpp in core/)
