+++ include/seqan3/range/container/bitcompressed_vector.hpp
@@ -17,7 +17,7 @@
 #include <sdsl/int_vector.hpp>
 
 #include <seqan3/alphabet/detail/alphabet_proxy.hpp>
-#include <seqan3/core/math.hpp>
+#include <seqan3/utility/math.hpp>
 #include <seqan3/core/concept/cereal.hpp>
 #include <seqan3/range/detail/random_access_iterator.hpp>
 #include <seqan3/range/views/to_char.hpp>
