+++ test/unit/core/CMakeLists.txt
@@ -3,7 +3,6 @@ add_subdirectories()
 seqan3_test (add_enum_bitwise_operators_test.cpp)
 seqan3_test (bit_manipulation_test.cpp)
 seqan3_test (debug_stream_test.cpp)
-seqan3_test (math_test.cpp)
 seqan3_test (pod_tuple_test.cpp)
 seqan3_test (tuple_utility_test.cpp)
 seqan3_test (type_list_test.cpp)
