+++ test/unit/utility/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectories()
+
+seqan3_test (math_test.cpp)
